requirements.txt
langchain
langchain-community
deeplake
openai
tiktoken
gitpython
python-dotenv
streamlit
psycopg2-binary==2.9.9